refactor(SentinelOne): move the filter_collected_events method upper in the module
