+++ SentinelOne/sentinelone_module/logs/helpers.py
@@ -1,7 +1,5 @@
 from datetime import datetime
-from typing import Callable, Sequence
 
-from cachetools import Cache
 from management.mgmtsdk_v2.entities.activity import Activity
 from management.mgmtsdk_v2.entities.threat import Threat
 
@@ -27,27 +25,3 @@ def get_latest_event_timestamp(events: list[Activity | Threat | dict]) -> dateti
                     latest_event_datetime = event_created_at
 
     return latest_event_datetime
-
-
-def filter_collected_events(events: Sequence, getter: Callable, cache: Cache) -> list:
-    """
-    Filter events that have already been filter_collected_events
-
-    Args:
-        events: The list of events to filter
-        getter: The callable to get the criteria to filter the events
-        cache: The cache that hold the list of collected events
-    """
-
-    selected_events = []
-    for event in events:
-        key = getter(event)
-
-        # If the event was already collected, discard it
-        if key is None or key in cache:
-            continue
-
-        cache[key] = True
-        selected_events.append(event)
-
-    return selected_events
