+++ SentinelOne/tests/logs/test_helpers.py
@@ -1,11 +1,10 @@
 from datetime import datetime, timezone
 
 import pytest
-from cachetools import LRUCache
 from management.mgmtsdk_v2.entities.activity import Activity
 from management.mgmtsdk_v2.entities.threat import Threat
 
-from sentinelone_module.logs.helpers import filter_collected_events, get_latest_event_timestamp
+from sentinelone_module.logs.helpers import get_latest_event_timestamp
 
 
 @pytest.mark.parametrize(
@@ -41,33 +40,3 @@ from sentinelone_module.logs.helpers import filter_collected_events, get_latest_
 )
 def test_get_lastest_event_timestamp(events, expected_datetime):
     assert get_latest_event_timestamp(events) == expected_datetime
-
-
-@pytest.mark.parametrize(
-    "events,getter,cache,expected_list",
-    [
-        (
-            ["key1", "key2", "key3", "key2", "key4"],
-            lambda x: x,
-            LRUCache(maxsize=256),
-            ["key1", "key2", "key3", "key4"],
-        ),
-        ([], lambda x: x, LRUCache(maxsize=256), []),
-        (["key1", "key2", "key3", "key4"], lambda x: x, LRUCache(maxsize=256), ["key1", "key2", "key3", "key4"]),
-        (
-            [
-                {"name": "key1"},
-                {"name": "key2"},
-                {"name": "key3"},
-                {"name": "key1"},
-                {"name": "key4"},
-                {"key": "key5"},
-            ],
-            lambda x: x.get("name"),
-            LRUCache(maxsize=256),
-            [{"name": "key1"}, {"name": "key2"}, {"name": "key3"}, {"name": "key4"}],
-        ),
-    ],
-)
-def test_filter_collected_events(events, getter, cache, expected_list):
-    assert filter_collected_events(events, getter, cache) == expected_list
